+++ JavaCompiler/Util.cpp
@@ -18,7 +18,7 @@ namespace Util {
   std::string trim(const std::string str) {
     int start, end;
     for(start = 0; start < str.length() && isWhiteSpace(str[start]); start++);
-    for(end = str.length() - 1; end >= 0 && isWhiteSpace(str[end]); end--);
+    for(end = (int)str.length() - 1; end >= 0 && isWhiteSpace(str[end]); end--);
     if(end <= start) return EMPTY_STRING;
     return str.substr(start, end - start + 1);
   }
